Allocate leave and update allocation count

+++ hrm-commons/src/main/java/com/sgic/hrm/commons/dto/LeaveTypeData.java
@@ -4,50 +4,50 @@ import java.time.ZonedDateTime;
 
 public class LeaveTypeData {
 
-	private Integer id;
-	private String leaveTypeValue;
-	private Float defaultAllocation;
-	private ZonedDateTime createdAt;
-	private ZonedDateTime updatedAt;
+  private Integer id;
+  private String leaveTypeValue;
+  private Double defaultAllocation;
+  private ZonedDateTime createdAt;
+  private ZonedDateTime updatedAt;
 
-	public Integer getId() {
-		return id;
-	}
+  public Integer getId() {
+    return id;
+  }
 
-	public void setId(Integer id) {
-		this.id = id;
-	}
+  public void setId(Integer id) {
+    this.id = id;
+  }
 
-	public String getLeaveTypeValue() {
-		return leaveTypeValue;
-	}
+  public String getLeaveTypeValue() {
+    return leaveTypeValue;
+  }
 
-	public void setLeaveTypeValue(String leaveTypeValue) {
-		this.leaveTypeValue = leaveTypeValue;
-	}
+  public void setLeaveTypeValue(String leaveTypeValue) {
+    this.leaveTypeValue = leaveTypeValue;
+  }
 
-	public Float getDefaultAllocation() {
-		return defaultAllocation;
-	}
+  public Double getDefaultAllocation() {
+    return defaultAllocation;
+  }
 
-	public void setDefaultAllocation(Float defaultAllocation) {
-		this.defaultAllocation = defaultAllocation;
-	}
+  public void setDefaultAllocation(Double defaultAllocation) {
+    this.defaultAllocation = defaultAllocation;
+  }
 
-	public ZonedDateTime getCreatedAt() {
-		return createdAt;
-	}
+  public ZonedDateTime getCreatedAt() {
+    return createdAt;
+  }
 
-	public void setCreatedAt(ZonedDateTime createdAt) {
-		this.createdAt = createdAt;
-	}
+  public void setCreatedAt(ZonedDateTime createdAt) {
+    this.createdAt = createdAt;
+  }
 
-	public ZonedDateTime getUpdatedAt() {
-		return updatedAt;
-	}
+  public ZonedDateTime getUpdatedAt() {
+    return updatedAt;
+  }
 
-	public void setUpdatedAt(ZonedDateTime updatedAt) {
-		this.updatedAt = updatedAt;
-	}
+  public void setUpdatedAt(ZonedDateTime updatedAt) {
+    this.updatedAt = updatedAt;
+  }
 
 }
